+++ CHANGELOG.md
@@ -164,3 +164,22 @@
 ### Version 101.0.2 :
 * Add documentation link in the connector admin configuration page
 * Fix metric attribute unit missing test case
+
+### Version 101.1.0 :
+* Add new Akeneo Edition selector configuration (https://help.akeneo.com/magento2-connector/v100/articles/02-configure-PIM-API.html#configure-your-akeneo-edition)
+
+**Warning :** *After updating Akeneo Connector for Magento 2 to this version, make sure to configure the correct Akeneo Edition in your connector configuration*
+
+* Add category tree filtering from API for Akeneo version 4.0.62 or greater and Akeneo Serenity
+* Add family updated date filter in connector configuration for Akeneo version 4.0.62 or greater and Akeneo Serenity (https://help.akeneo.com/magento2-connector/v100/articles/10-configure-families.html#how-can-i-filter-my-families-during-import)
+
+### Version 101.1.1 :
+* Add attribute filtering by type from API in the admin configuration page for Akeneo version 4.0.62 or greater and Akeneo Serenity
+* Add attribute updated date filter in connector configuration for Akeneo version 4.0.62 or greater and Akeneo Serenity (https://help.akeneo.com/magento2-connector/v100/articles/11-filter-attributes.html)
+* Add attribute filter by code in connector configuration for Akeneo version 4.0.62 or greater and Akeneo Serenity (https://help.akeneo.com/magento2-connector/v100/articles/11-filter-attributes.html)
+* Fix category tree filtering when no categories are excluded for Akeneo version 4.0.62 or greater and Akeneo Serenity
+* Fix attribute job when no attribute is found with correct label
+* Fix metric option import when no option is found
+
+### Version 101.1.2 :
+* Fix image attribute import to fill the catalog_product_entity_media_gallery_value table